+++ lib/domain/omh/bloodpressure.g.dart
@@ -21,10 +21,7 @@ BloodPressure _$BloodPressureFromJson(Map<String, dynamic> json) {
         ? null
         : TimeFrame.fromJson(
             json['effective_time_frame'] as Map<String, dynamic>)
-    ..descriptiveStatistic = json['descriptive_statistic'] == null
-        ? null
-        : DescriptiveStatistic.fromJson(
-            json['descriptive_statistic'] as Map<String, dynamic>)
+    ..descriptiveStatistic = json['descriptive_statistic'] as String
     ..userNotes = json['user_notes'] as String;
 }
 
@@ -84,41 +81,5 @@ Map<String, dynamic> _$DiastolicBloodPressureToJson(
 
   writeNotNull('unit', instance.unit);
   writeNotNull('value', instance.value);
-  return val;
-}
-
-BloodPressureUnit _$BloodPressureUnitFromJson(Map<String, dynamic> json) {
-  return BloodPressureUnit(json['schema_value']);
-}
-
-Map<String, dynamic> _$BloodPressureUnitToJson(BloodPressureUnit instance) {
-  var val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('schema_value', instance.schemaValue);
-  return val;
-}
-
-PositionDuringMeasurement _$PositionDuringMeasurementFromJson(
-    Map<String, dynamic> json) {
-  return PositionDuringMeasurement(json['schema_value']);
-}
-
-Map<String, dynamic> _$PositionDuringMeasurementToJson(
-    PositionDuringMeasurement instance) {
-  var val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('schema_value', instance.schemaValue);
   return val;
 }