+++ lib/domain/omh/measure.g.dart
@@ -12,10 +12,7 @@ Measure _$MeasureFromJson(Map<String, dynamic> json) {
         ? null
         : TimeFrame.fromJson(
             json['effective_time_frame'] as Map<String, dynamic>)
-    ..descriptiveStatistic = json['descriptive_statistic'] == null
-        ? null
-        : DescriptiveStatistic.fromJson(
-            json['descriptive_statistic'] as Map<String, dynamic>)
+    ..descriptiveStatistic = json['descriptive_statistic'] as String
     ..userNotes = json['user_notes'] as String;
 }