+++ lib/openmhealth_schemas.dart
@@ -9,3 +9,4 @@ export 'package:openmhealth_schemas/domain/omh/geoposition.dart';
 export 'package:openmhealth_schemas/domain/omh/activity.dart';
 export 'package:openmhealth_schemas/domain/omh/bloodpressure.dart';
 export 'package:openmhealth_schemas/domain/omh/datapoint.dart';
+export 'package:openmhealth_schemas/domain/omh/acceleration.dart';